added a function to look up set/binding indices for a descriptor name
The lookup is kind of painful and verbose, but I don't WANT it to be easy;
this encourages callers to cache the results.

Some descriptors (like uniform buffers) don't require names, in which
their contents are added to the global namespace. Looking them up is...
impossible, I guess, because the block name doesn't make it into SPIRV.
For now I just named everything, which seems a decent practice anyway.