First pass at handling window iconify/restore
Previously, an iconified (minimized) window would report its size as 0x0,
which would cause a validation error during swapchain recreation. Now,
if we detect an iconified window at the beginning of the main loop, we
just skip the entire loop body and sleep for 17ms. When the window is
restored, the main loop continues (possible recreating the swapchain if
the dimensions have changed, but otherwise keeping it as-is).

Just to be extra careful, code was added to early-out of acquire or present
if the framebuffer size is zero, but I don't *think* that should be
necessary; those values should be stable between calls to glfwPollEvents().